Internals

Auto generated docs for functions and types that are not exported. Mostly for development use.

Types

SpinAdaptedSecondQuantization.Constraints — Type

Constraints = SortedDict{Int,Type}

Type alias for container of MO-Index constraints

SpinAdaptedSecondQuantization.KroneckerDelta — Type

KroneckerDelta

Type representing a Kronecker delta of two or more MO-indices.

SpinAdaptedSecondQuantization.Operator — Type

Operator

Abstact operator type which all concrete operator types (i.e. Epq, a†p) must extend.

SpinAdaptedSecondQuantization.SingletExcitationOperator — Type

SingletExcitationOperator

The basic E_pq type operator.

SpinAdaptedSecondQuantization.Spin — Type

Spin

Enum type to represent spin up (α) and down (β)

Functions

SpinAdaptedSecondQuantization.disable_color — Method

disable_color()

Disables the coloring of MO-indices to indicate orbital space constraints. This will make all constraints be printed explicitly which can make some terms a bit long to read. Color is enabled by default.

SpinAdaptedSecondQuantization.enable_color — Method

enable_color()

Enables the coloring of MO-indices to indicate orbital space constraints. Color is enabled by default. Currently the coloring scheme is given by: GeneralOrbital: colorless OccupiedOrbital: green VirtualOrbital: blue