Reference/API

Auto generated docs for all exported functions and types.

Types

Functions

SpinAdaptedSecondQuantization.E — Method

E(p, q)

Constructs an expression containing a single excitation operator.

SpinAdaptedSecondQuantization.act_on_ket — Method

act_on_ket(::Operator)

All typed extending Operator must overload this function. It should return the expression you end up with after acting the operator on a normal RHF ket.

Example:

$E_{p q} |HF⟩ = (2 δ_{p q} C(p∈O,q∈O) + E_{p q} C(p∈V,q∈O)) |HF⟩$

SpinAdaptedSecondQuantization.delta — Method

delta(indices...)

Construct an expression consisting of a single Kronecker delta of the given indices.

SpinAdaptedSecondQuantization.disable_color — Method

disable_color(S)

Disables the coloring for a specific orbital space. This can be useful when dealing with a medium-large number of orbital spaces, and having distinguishable colors is infeasible.

SpinAdaptedSecondQuantization.disable_color — Method

disable_color()

Disables the coloring of MO-indices to indicate orbital space constraints globally. This will make all constraints be printed explicitly which can make some terms a bit long to read. Color is enabled by default.

SpinAdaptedSecondQuantization.disable_index_translation — Method

disables_index_translation()

Disables the translation of summation indices over occupied and virtual to be printed as ijkl... and abcd... respectively instead of pqrs...

See enable_index_translation

SpinAdaptedSecondQuantization.e — Method

e(p, q, r, s) = E(p, q) * E(r, s) - δ(r, q) * E(p, s)

Alias for the two electron singlet excitation operator. ```

SpinAdaptedSecondQuantization.enable_color — Method

enable_color()

Enables the coloring of MO-indices to indicate orbital space constraints. Color is enabled by default. The default coloring scheme is given by:

GeneralOrbital: :nothing

OccupiedOrbital: :light_green

VirtualOrbital: :cyan

SpinAdaptedSecondQuantization.enable_index_translation — Method

enable_index_translation()

Enables the translation of summation indices over occupied and virtual to be printed as ijkl... and abcd... respectively instead of pqrs...

By default translated indices lose their subspace coloring to reduce redundant information (unless the index is in an even stricter subspace which requires coloring nevertheless).

This is enabled by default.

SpinAdaptedSecondQuantization.print_code — Method

Print term to numpy's einsum routine

SpinAdaptedSecondQuantization.print_eT_code — Method

Used with https://github.com/alexancp/einsumpath-to-eT to generate Fortran subroutines for eT.

SpinAdaptedSecondQuantization.print_eT_function_generator — Function

printeTfunctiongenerator(name, ex, symbol, indices, translation, [wftype], [tensortranslation], [noinputtensors], [outperms])

SpinAdaptedSecondQuantization.print_julia_function — Function

printjuliafunction(name, ex, symbol, indices, translation, [tensortranslation], [explicittensor_blocks])

SpinAdaptedSecondQuantization.project_biorthogonal — Method

project_biorthogonal(ex::Expression, template::Expression)

This has the effect of projecting the expression ex on the biorthogonal bra state of the template expression.

For a singlet double excitation, the following will be equivalent:

bra = (
        1//3 * E(1, 2) * E(3, 4) +
        1//6 * E(1, 4) * E(3, 2)
    )' * occupied(2, 4) * virtual(1, 3)

result = hf_expectation_value(bra * ex)

ket = E(1, 2) * E(3, 4) * occupied(2, 4) * virtual(1, 3)
result = project_biorthogonal(act_on_ket(ex), ket)
result = symmetrize(result, make_permutation_mappings([(1, 2), (3, 4)]))

SpinAdaptedSecondQuantization.set_color — Method

set_color(S, color=default_color(S))

Sets the color of a given orbital space to the specified color, or the default if none is given. Allowed colors are given by the Base.text_colors dict. This includes integers in the range 0-255 which the corresponding colors can be found in this table on Wikipedia.

SpinAdaptedSecondQuantization.δ — Method

δ(indices...)

Unicode alias for delta.

SpinAdaptedSecondQuantization.τ — Method

τ(p, q)

Constructs an expression containing a single triplet excitation operator.