Reference/API

Auto generated docs for all exported functions and types.

Types

Functions

SpinAdaptedSecondQuantization.E — Method

E(p, q)

Constructs an expression containing a single excitation operator.

SpinAdaptedSecondQuantization.act_eT_on_bra — Method

act_eT_on_bra(braop::Expression, T::Expression; [max_n=Inf], [max_ops=Inf])

computes

$\langle\text{HF}| \text{braop} e^T$

This can be useful for evaluating certain coupled cluster expressions from left to right. Since the cluster operator T is purely exciting acting $e^T$ to the left will terminate after only a few terms depending on the left state being projected on.

Warning

If T contains de-excitations, the expression will not truncate and the parameter max_n must be specified with a finite positive integer for the function to terminate, truncating the expression. max_n specifies the highest order term from the taylor expansion of e^T to include. If T is purely exciting, the expansion exits early when it terminates.

SpinAdaptedSecondQuantization.act_on_bra — Function

act_on_bra(x::Expression, [max_ops=Inf])

Projects x on a HF bra by projecting the adjoint on a ket and ajointing the result.

act_on_bra(x, max_ops=Inf) = act_on_ket(x', max_ops)'

See act_on_ket

SpinAdaptedSecondQuantization.act_on_ket — Method

act_on_ket(::Operator)

All typed extending Operator must overload this function. It should return the expression you end up with after acting the operator on a normal RHF ket.

Example:

$E_{p q} |HF⟩ = (2 δ_{p q} C(p∈O,q∈O) + E_{p q} C(p∈V,q∈O)) |HF⟩$

SpinAdaptedSecondQuantization.anticommutator — Method

anticommutator(a::Expression, b::Expression)

Compute commutator between a and b. This is computed termwise using the reductive_commutator function, then adding/subtracting the swapped result if the reductive commutator for a given pair of terms turned out to be an commutator.

SpinAdaptedSecondQuantization.bch — Method

bch(A, Bs::AbstractArray, n)

Equivalent to bch(A, sum(Bs), n) if all Bs commute with eachother, but evalutating in a smart way in order to avoid computing equivalent terms many times, for example instead of computing both $rac12 [[A, B_1], B_2]$ and $rac12 [[A, B_2], B_1]$, it will only compute the first one and multiply by 2. This can be much faster if one has many commuting terms in Bs.

Warning

Should not be used if some expressions in Bs do not commute.

SpinAdaptedSecondQuantization.bch — Method

bch(A, B, n)

Compute Baker-Campbell-Haussdorff expansion

$e^{-B} A e^B = A + [A, B] + \frac{1}{2!} [[A, B], B] + \frac{1}{3!} [[[A, B], B], B] + \frac{1}{n!} [...[[[A, B], B], B], ...B]$

up to nth order. Terminates if commutator becomes zero.

SpinAdaptedSecondQuantization.commutator — Method

commutator(A, B, C, ...)

Computes the nested commutator $[...[[A, B], C], ...]$

SpinAdaptedSecondQuantization.commutator — Method

commutator(A::Expression, B::Expression, n::Integer)

Compute n nested commutators $[...[[[A], B], B]..., B]$.

SpinAdaptedSecondQuantization.commutator — Method

commutator(a::Expression, b::Expression)

Compute commutator between a and b. This is computed termwise using the reductive_commutator function, then adding/subtracting the swapped result if the reductive commutator for a given pair of terms turned out to be an anti-commutator.

SpinAdaptedSecondQuantization.convert_to_elementary_operators — Method

convert_to_elementary_operators(ex::Expression)

converts all operators to their representation using "elementary" annihilation and creation operators if implemented for all operator types in ex.

SpinAdaptedSecondQuantization.delta — Method

delta(indices...)

Construct an expression consisting of a single Kronecker delta of the given indices.

SpinAdaptedSecondQuantization.desymmetrize — Method

desymmetrize(ex::Expression, mappings)

SpinAdaptedSecondQuantization.disable_color — Method

disable_color(S)

Disables the coloring for a specific orbital space. This can be useful when dealing with a medium-large number of orbital spaces, and having distinguishable colors is infeasible.

SpinAdaptedSecondQuantization.disable_color — Method

disable_color()

Disables the coloring of MO-indices to indicate orbital space constraints globally. This will make all constraints be printed explicitly which can make some terms a bit long to read. Color is enabled by default.

SpinAdaptedSecondQuantization.disable_index_translation — Method

disables_index_translation()

Disables the translation of summation indices over occupied and virtual to be printed as ijkl... and abcd... respectively instead of pqrs...

See enable_index_translation

SpinAdaptedSecondQuantization.e — Method

e(p, q, r, s) = E(p, q) * E(r, s) - δ(r, q) * E(p, s)

Alias for the two electron singlet excitation operator. ```

SpinAdaptedSecondQuantization.enable_color — Method

enable_color()

Enables the coloring of MO-indices to indicate orbital space constraints. Color is enabled by default. The default coloring scheme is given by:

GeneralOrbital: :nothing

OccupiedOrbital: :light_green

VirtualOrbital: :cyan

SpinAdaptedSecondQuantization.enable_index_translation — Method

enable_index_translation()

Enables the translation of summation indices over occupied and virtual to be printed as ijkl... and abcd... respectively instead of pqrs...

By default translated indices lose their subspace coloring to reduce redundant information (unless the index is in an even stricter subspace which requires coloring nevertheless).

This is enabled by default.

SpinAdaptedSecondQuantization.look_for_tensor_replacements — Method

look_for_tensor_replacements(ex::Expression, transformer)

Looks for pairs of terms in ex that are equal up to a tensor transformation given by transformer. The most common use case is to recognize patterns such as $L_{pqrs} = 2 g_{pqrs} - g_{psrq}$ where a convencience function to produce the relevant transformer provided (see make_exchange_transformer).

DocTestSetup = quote
    using SpinAdaptedSecondQuantization
    h = ∑(real_tensor("h", 1, 2) * E(1, 2) * electron(1, 2), 1:2)
    g = 1//2 * simplify(
        ∑(psym_tensor("g", 1:4...) * e(1:4...) * electron(1:4...), 1:4)
    )
    H = h + g + real_tensor("h_nuc")
end

julia> E_HF = simplify_heavy(hf_expectation_value(H))
h_nuc
+ 2 ∑_i(h_ii)
+ 2 ∑_ij(g_iijj)
- ∑_ij(g_ijji)
julia> look_for_tensor_replacements(E_HF, make_exchange_transformer("g", "L"))
h_nuc
+ 2 ∑_i(h_ii)
+ ∑_ij(L_iijj)

SpinAdaptedSecondQuantization.make_exchange_transformer — Method

make_exchange_transformer(name_from, name_to)

Convenience function to make tensor transformer for use in look_for_tensor_replacements.

SpinAdaptedSecondQuantization.make_permutation_mappings — Method

make_permutation_mappings(index_groups)

Convenience function to make mapping for all permutations of groups of indices. See symmetrize and desymmetrize.

SpinAdaptedSecondQuantization.print_code — Method

Print term to numpy's einsum routine

SpinAdaptedSecondQuantization.print_eT_code — Method

Used with https://github.com/alexancp/einsumpath-to-eT to generate Fortran subroutines for eT.

SpinAdaptedSecondQuantization.print_eT_function_generator — Function

printeTfunctiongenerator(name, ex, symbol, indices, translation, [wftype], [tensortranslation], [noinputtensors], [outperms])

SpinAdaptedSecondQuantization.print_julia_function — Function

printjuliafunction(name, ex, symbol, indices, translation, [tensortranslation], [explicittensor_blocks])

SpinAdaptedSecondQuantization.project_biorthogonal — Method

project_biorthogonal(ex::Expression, template::Expression)

This has the effect of projecting the expression ex on the biorthogonal bra state of the template expression.

For a singlet double excitation, the following will be equivalent:

bra = (
        1//3 * E(1, 2) * E(3, 4) +
        1//6 * E(1, 4) * E(3, 2)
    )' * occupied(2, 4) * virtual(1, 3)

result = hf_expectation_value(bra * ex)

ket = E(1, 2) * E(3, 4) * occupied(2, 4) * virtual(1, 3)
result = project_biorthogonal(act_on_ket(ex), ket)
result = symmetrize(result, make_permutation_mappings([(1, 2), (3, 4)]))

SpinAdaptedSecondQuantization.set_color — Method

set_color(S, color=default_color(S))

Sets the color of a given orbital space to the specified color, or the default if none is given. Allowed colors are given by the Base.text_colors dict. This includes integers in the range 0-255 which the corresponding colors can be found in this table on Wikipedia.

SpinAdaptedSecondQuantization.simplify — Method

simplify(ex::Expression)

Applies various simplification strategies to the given expression. The most important are:

Remove Kronecker deltas when indices are summed over.

Perform various operations to try to lower the lexiographic ordering of the each term including:

Lowering indices showing up in a Kronecker delta to the lowest of the indices in said delta.

Lowering summation indices to lowest available indices.

Reorder summation indices to lower the lexiographic ordering of the order they show up within the term.

Bubble sort operators swaping neighbouring operators if:

They commute

The swap would lower the terms lexiographic ordering.

Look for pairs of terms differing only in index constraints and scalar prefactors, and tries to combine them.

Examples:

Collapsing deltas:

julia> using SpinAdaptedSecondQuantization

julia> ∑(real_tensor("h", 1, 2) * electron(1, 2) * δ(1, 2), [2])
∑_q(δ_pq h_pq)
julia> simplify(ans)
h_pp

Combining terms differing only by index constraints and scalars:

julia> real_tensor("h", 1, 2) * (occupied(1, 2) + occupied(1) * virtual(2))
h_ia
+ h_ij
julia> simplify(ans)
h_ip

julia> ∑(real_tensor("h", 1, 2) * E(1, 2) * electron(1, 2), [1, 2])
∑_pq(h_pq E_pq)
julia> ans - ∑(real_tensor("h", 1, 2) * E(1, 2) * electron(1) * occupied(2), [1, 2])
∑_pq(h_pq E_pq)
- ∑_pi(h_pi E_pi)
julia> simplify(ans)
∑_pa(h_pa E_pa)

Sorting commuting operators:

julia> E(3, 4) * E(1, 2) * electron(1:4...)
E_rs E_pq
julia> simplify(ans) # not able to swap as commutator is not zero
E_rs E_pq
julia> ans * occupied(2, 4) * virtual(1, 3)
E_bj E_ai
julia> simplify(ans) # now able to swap
E_ai E_bj

julia> E(3, 4) * E(5, 6) * E(1, 2) * electron(5, 6) * occupied(2, 4) * virtual(1, 3)
E_bj E_pq E_ai
julia> simplify(ans) # Can not swap first and last because of non commuting between
E_bj E_pq E_ai
julia> E(5, 6) * E(3, 4) * E(1, 2) * electron(5, 6) * occupied(2, 4) * virtual(1, 3)
E_pq E_bj E_ai
julia> simplify(ans) # Can swap adjacent last two, but not first
E_pq E_ai E_bj

SpinAdaptedSecondQuantization.simplify_heavy — Function

simplify_heavy(ex::Expression,
    [mapping=GeneralOrbital => (OccupiedOrbital, VirtualOrbital)])

Similar to simplify, but enabling a few extra simplification strategies that can be quite expensive, especially as the number of summation indices in a term grows large.

The most expensive step is to iterate over every permutation of the summation indices of each term in order to find the order that has the least lexiographic ordering, which naturally has a scaling of O(n!) where n is the larget number of summation indices in a single term. This is quite managable for "small" numbers of indices (less than ~8), but quickly becomes unfeasible.

Note

The iteration over permutations of summation indices happens after summation indices that show up in Kronecker deltas, so if your terms contain many Kronecker deltas this might yet be possible to run.

The optional mapping argument defaults to splitting terms with general indices into a term with the index being occupied plus a term where the index is virtual, then performing simplification when all terms are split and recombining terms according to the same splitting if possible. This often allows for a few additional cancellations to occur than using only simplify If dealing with other index spaces, this splitting can be specified can be specified to be some other set of index spaces. See split_indices.

Examples:

HF energy with full real orbital symmetry:

julia> using SpinAdaptedSecondQuantization

julia> h = ∑(rsym_tensor("h", 1, 2) * E(1, 2) * electron(1, 2), 1:2)
∑_pq(h_pq E_pq)
julia> g = 1//2 * ∑(rsym_tensor("g", 1:4...) * e(1:4...) * electron(1:4...), 1:4)
1/2 ∑_pqrs(g_pqrs E_pq E_rs)
- 1/2 ∑_pqrs(δ_qr g_pqrs E_ps)
julia> H = simplify(h + g + real_tensor("h_nuc"));

julia> E_HF = simplify(hf_expectation_value(H)) # Not nice
h_nuc
+ 2 ∑_i(h_ii)
+ 2 ∑_ij(g_iijj)
- ∑_pi(g_pipi)
+ ∑_ia(g_iaia)
julia> simplify_heavy(E_HF) # Much nicer!
h_nuc
+ 2 ∑_i(h_ii)
+ 2 ∑_ij(g_iijj)
- ∑_ij(g_ijij)

Simplification by combining terms differing by index constraints

julia> ∑(real_tensor("h", 1, 2) * E(1, 2) * electron(1, 2), [1, 2])
∑_pq(h_pq E_pq)
julia> ans - ∑(real_tensor("h", 1, 2) * E(1, 2) * occupied(1, 2), [1, 2])
∑_pq(h_pq E_pq)
- ∑_ij(h_ij E_ij)
julia> ans - ∑(real_tensor("h", 1, 2) * E(1, 2) * virtual(1, 2), [1, 2])
∑_pq(h_pq E_pq)
- ∑_ab(h_ab E_ab)
- ∑_ij(h_ij E_ij)
julia> simplify(ans)
∑_pq(h_pq E_pq)
- ∑_ab(h_ab E_ab)
- ∑_ij(h_ij E_ij)
julia> simplify_heavy(ans)
∑_ai(h_ai E_ai)
+ ∑_ia(h_ia E_ia)

SpinAdaptedSecondQuantization.symmetrize — Method

symmetrize(ex::Expression, mappings)

Function to expand the expression ex including all permutations of indices given by mappings. This is common to use to expand all permuations among neighbouring pairs of indices, such as to compute $P_{ijk}^{abc} \Omega_{ijk}^{abc}$. For this specific purpose we provide the convenience function make_permutation_mappings which produces all permutations between groups of indices.

DocTestSetup = :(using SpinAdaptedSecondQuantization)

julia> x = real_tensor("t", 1, 2, 3, 4, 5, 6) *
           occupied(2, 4, 6) * virtual(1, 3, 5)
t_aibjck
julia> symmetrize(x, make_permutation_mappings([(1, 2), (3, 4), (5, 6)]))
t_aibjck
+ t_aickbj
+ t_bjaick
+ t_bjckai
+ t_ckaibj
+ t_ckbjai

SpinAdaptedSecondQuantization.δ — Method

δ(indices...)

Unicode alias for delta.

SpinAdaptedSecondQuantization.τ — Method

τ(p, q)

Constructs an expression containing a single triplet excitation operator.