Reference/API

Auto generated docs for all exported functions and types.

Types

Functions

SpinAdaptedSecondQuantization.E — Method

E(p, q)

Constructs an expression containing a single excitation operator.

SpinAdaptedSecondQuantization.act_on_ket — Method

act_on_ket(::Operator)

All typed extending Operator must overload this function. It should return the expression you end up with after acting the operator on a normal RHF ket.

Example:

$E_{p q} |HF⟩ = (2 δ_{p q} C(p∈O,q∈O) + E_{p q} C(p∈V,q∈O)) |HF⟩$

SpinAdaptedSecondQuantization.delta — Method

delta(indices...)

Construct an expression consisting of a single Kronecker delta of the given indices.

SpinAdaptedSecondQuantization.e — Method

e(p, q, r, s) = E(p, q) * E(r, s) - δ(r, q) * E(p, s)

Alias for the two electron singlet excitation operator. ```

SpinAdaptedSecondQuantization.δ — Method

δ(indices...)

Unicode alias for delta.

OrbitalSpaces

SpinAdaptedSecondQuantization.GeneralOrbital — Type

GeneralOrbital

Top level supertype of all orbital spaces. Any orbital subspace is a subtype of this type.

SpinAdaptedSecondQuantization.OccupiedOrbital — Type

OccupiedOrbital

Type representing all occupied orbitals.

SpinAdaptedSecondQuantization.VirtualOrbital — Type

VirtualOrbital

Type representing all virtual orbitals.