Reference/API

Auto generated docs for all exported functions and types.

Types

Functions

SpinAdaptedSecondQuantization.E — Method

E(p, q)

Constructs an expression containing a single excitation operator.

SpinAdaptedSecondQuantization.act_on_ket — Method

act_on_ket(::Operator)

All typed extending Operator must overload this function. It should return the expression you end up with after acting the operator on a normal RHF ket.

Example:

$E_{p q} |HF⟩ = (2 δ_{p q} C(p∈O,q∈O) + E_{p q} C(p∈V,q∈O)) |HF⟩$

SpinAdaptedSecondQuantization.delta — Method

delta(indices...)

Construct an expression consisting of a single Kronecker delta of the given indices.

SpinAdaptedSecondQuantization.disable_color — Method

disable_color()

Disables the coloring of MO-indices to indicate orbital space constraints globally. This will make all constraints be printed explicitly which can make some terms a bit long to read. Color is enabled by default.

SpinAdaptedSecondQuantization.disable_color — Method

disable_color(S)

Disables the coloring for a specific orbital space. This can be useful when dealing with a medium-large number of orbital spaces, and having distinguishable colors is infeasible.

SpinAdaptedSecondQuantization.disable_color_translated — Method

disable_color_translated()

Disables the coloring of indices that has been translated.

SpinAdaptedSecondQuantization.disable_index_translation — Method

disables_index_translation()

Disables the translation of summation indices over occupied and virtual to be printed as ijkl... and abcd... respectively instead of pqrs...

See enable_index_tranlation

SpinAdaptedSecondQuantization.e — Method

e(p, q, r, s) = E(p, q) * E(r, s) - δ(r, q) * E(p, s)

Alias for the two electron singlet excitation operator. ```

SpinAdaptedSecondQuantization.enable_color — Method

enable_color()

Enables the coloring of MO-indices to indicate orbital space constraints. Color is enabled by default. The default coloring scheme is given by:

GeneralOrbital: :nothing

OccupiedOrbital: :light_green

VirtualOrbital: :cyan

SpinAdaptedSecondQuantization.enable_color_translated — Method

enable_color_translated()

Enables the coloring of indices that has been translated.

SpinAdaptedSecondQuantization.enable_index_tranlation — Method

enable_index_translation()

Enables the translation of summation indices over occupied and virtual to be printed as ijkl... and abcd... respectively instead of pqrs...

By default translated indices lose their subspace coloring to reduce redundant information (unless the index is in an even stricter subspace which requires coloring nevertheless). To re-enable the coloring see enable_color_translated.

This is enabled by default.

SpinAdaptedSecondQuantization.print_code — Method

Print term to numpy's einsum routine

SpinAdaptedSecondQuantization.print_eT_code — Method

Used with https://github.com/alexancp/einsumpath-to-eT to generate Fortran subroutines for eT.

SpinAdaptedSecondQuantization.set_color — Method

set_color(S, [color=default_color(S)])

Sets the color of a given orbital space to the specified color, or the default if none is given. Allowed colors are given by the Base.text_colors dict. This includes integers in the range 0-255 which the corresponding colors can be found in this table on Wikipedia.

SpinAdaptedSecondQuantization.δ — Method

δ(indices...)

Unicode alias for delta.

SpinAdaptedSecondQuantization.τ — Method

τ(p, q)

Constructs an expression containing a single triplet excitation operator.

OrbitalSpaces

SpinAdaptedSecondQuantization.GeneralOrbital — Type

GeneralOrbital

Top level supertype of all orbital spaces. Any orbital subspace is a subtype of this type.

SpinAdaptedSecondQuantization.OccupiedOrbital — Type

OccupiedOrbital

Type representing all occupied orbitals.

SpinAdaptedSecondQuantization.VirtualOrbital — Type

VirtualOrbital

Type representing all virtual orbitals.